src/python/FoundationaLLMAgentPlugins/requirements.txt
aiohttp==3.11.11
azure-appconfiguration-provider==1.2.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-monitor-opentelemetry==1.6.2
azure-monitor-opentelemetry-exporter==1.0.0b28
azure-search-documents==11.6.0b9
azure-storage-blob==12.21.0
boto3==1.36.13
foundationallm==0.9.7b105
langchain==0.3.17
langchain-aws==0.2.12
langchain-experimental==0.3.4
langchain-google-vertexai==2.0.13
langchain-openai==0.3.3
langchain-azure-dynamic-sessions==0.2.0
langgraph==0.2.53
openai==1.61.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
pandas==2.2.2
pydantic==2.10.6
unidecode==1.3.8
wikipedia==1.4.0


pyodbc==5.2.0
